requirements.txt
Flask
pandas
numpy
matplotlib
requests
aiohttp
pathlib
pyyaml
jinja2
aiohttp_jinja2
